extract setup code for pretix pipeline test as well
